requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
cryptography==44.0.0
distlib==0.3.9
Django==5.1.5
django-allauth==65.3.1
django-ckeditor==6.7.2
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-jalali-date==1.1.3
django-js-asset==3.0.1
django-rosetta==0.10.1
djangorestframework==3.15.2
drf-nested-routers==0.94.1
filelock==3.16.1
idna==3.10
jalali_core==1.0.0
jdatetime==5.1.0
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pipenv==2024.4.0
platformdirs==4.3.6
polib==1.2.0
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.8.0
sqlparse==0.5.3
termcolor==2.4.0
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.28.1
